Strip whitespace from cookie in firmware update
When updating firmware, some devices return a cookie with a newline in.
This can result in the following error:

Invalid return character or leading space in header: Cookie

This is due to the header validation in the requests library, which
requires no newlines in headers.

This change fixes the issue by stripping whitespace from the cookie.

Change-Id: Ib2dab7beef0da415e8f2358b459ec1205bf688b4
Closes-Bug: #1859616
(cherry picked from commit 37afea2bbc5fdd17f69a094d81a060f9140ed24a)