Fix etag header test
Proliantutils has a pattern of testing internal calls inside of
the Sushy library. This has the unfortunate side effect of becoming
broken as bugs and related issues are fixed.

Fixes the check on a response which upstream sushy has discovered
has become sensitive on several vendors where eTags must be utilized
to ensure a reliable and correct response from the BMC.

Related-Bug: #1976555
Change-Id: I6b802fbb6a3f93f67269c141cad02b58ccd2d3ac