Fix tox.ini for Zed
Upper constraints resolution was set to the master branch, but it
should be the Zed branch in order to avoid eventual test failures
as tools like pip value working over the stated precise bounds in
the repository. The upper constraints is the maximum upper bounds,
and will prevent inconsistent/misleading test failures in the
future.

Change-Id: Ifae5ff18c2d0f6a46b4775e3cb416bc551ac128d